requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
supabase==2.0.3
pandas==2.1.0
psycopg2-binary==2.9.9
postgrest==0.13.0
postgrest-py==0.4.0